+++ packages/integration-sdk-runtime/test/util/fixtures.ts
@@ -1,0 +1,42 @@
+import { ExecutionHistory, IntegrationInstance } from "@jupiterone/integration-sdk-core/src";
+
+export const LOCAL_INTEGRATION_INSTANCE: IntegrationInstance = {
+  id: 'local-integration-instance',
+  accountId: 'Your account',
+  name: 'Local Integration',
+  integrationDefinitionId: 'local-integration-definition',
+  description: 'A generated integration instance for local execution',
+  config: {},
+};
+
+export const LOCAL_EXECUTION_HISTORY: ExecutionHistory = {
+  current: {
+    startedOn: 0,
+  },
+};
+
+// eslint-disable-next-line @typescript-eslint/no-empty-function
+const noop = () => {};
+
+export function createMockIntegrationLogger(overrides) {
+  return {
+    trace: noop,
+    debug: noop,
+    info: noop,
+    warn: noop,
+    error: noop,
+    fatal: noop,
+    child: () => createMockIntegrationLogger(overrides),
+    isHandledError: () => true,
+    stepStart: noop,
+    stepSuccess: noop,
+    stepFailure: noop,
+    synchronizationUploadStart: noop,
+    synchronizationUploadEnd: noop,
+    validationFailure: noop,
+    publishMetric: noop,
+    publishEvent: noop,
+    publishErrorEvent: noop,
+    ...overrides,
+  };
+}